move packets reading to packets.js, remove some dead code
