+++ src/client.js
@@ -5,7 +5,6 @@ var EventEmitter = require('events').EventEmitter
   , compressPacketBuffer = protocol.compressPacketBuffer
   , oldStylePacket = protocol.oldStylePacket
   , newStylePacket = protocol.newStylePacket
-  , parsePacket = protocol.parsePacket
   , parsePacketData = protocol.parsePacketData
   , parseNewStylePacket = protocol.parseNewStylePacket
   , packetIds = protocol.packetIds
